Combine Schedule with AddressBook
