--- src/main/java/seedu/address/model/Schedule.java
@@ -1,23 +1,0 @@
-package seedu.address.model;
-
-import javafx.collections.ObservableList;
-import seedu.address.model.event.Event;
-import seedu.address.model.event.UniqueEventList;
-
-/**
- * Schedule contains all events
- */
-public class Schedule implements ReadOnlySchedule {
-    private final UniqueEventList events;
-
-    {
-        events = new UniqueEventList();
-    }
-
-    public Schedule() {}
-
-    @Override
-    public ObservableList<Event> getEventList() {
-        return events.asObservableList();
-    }
-}
