+++ src/main/java/seedu/address/model/event/Event.java
@@ -3,7 +3,7 @@ package seedu.address.model.event;
 import java.util.Date;
 
 /**
- * Represents an Event in the Schedule.
+ * Represents an Event in the schedule.
  */
 public interface Event {
     public String getTitle();
